Move saved analysis control to separate file
